refactor: delete expectation_suite_anonymizer
